+++ Code/src/OE/Platform/OpenGL/GLIndexBuffer.cpp
@@ -2,7 +2,7 @@
 
 namespace OrbitEngine {	namespace Graphics {
 	GLIndexBuffer::GLIndexBuffer(const size_t size, const void* data)
-		: GLBuffer(GL_ELEMENT_ARRAY_BUFFER, GL_STATIC_DRAW, size, data)
+		: GLBuffer(GL_ELEMENT_ARRAY_BUFFER, GL_STATIC_DRAW, (unsigned int)size, data)
 	{
 	}
 } }
