+++ Code/src/OE/Platform/OpenGL/GLTexture.cpp
@@ -26,7 +26,7 @@ namespace OrbitEngine { namespace Graphics {
 
 		applyParameters();
 
-		unsigned int size = data.size();
+		unsigned int size = (unsigned int)data.size();
 
 		if (size == 0) {
 			setData(nullptr);
@@ -41,8 +41,8 @@ namespace OrbitEngine { namespace Graphics {
 		else {
 			if (properties.dimension == TEXTURE2DARRAY)
 				glTexStorage3D(m_Target, size, m_InternalFormat, properties.formatProperties.width, properties.formatProperties.height, size);
-			for (size_t i = 0; i < size; i++)
-				setData(data[i], i);
+			for (int i = 0; i < int(size); i++)
+				setData(data[i], (unsigned int)i);
 		}
 
 		if (m_Properties.formatProperties.mipmapping)
