+++ Code/src/OE/Platform/OpenGL/GLVertexBuffer.cpp
@@ -2,7 +2,7 @@
 
 namespace OrbitEngine {	namespace Graphics {
 	GLVertexBuffer::GLVertexBuffer(const size_t size, const void* data)
-		: GLBuffer(GL_ARRAY_BUFFER, GL_DYNAMIC_DRAW, size, data)
+		: GLBuffer(GL_ARRAY_BUFFER, GL_DYNAMIC_DRAW, (unsigned int)size, data)
 	{
 	}
 } }
